Update UI and functionality to detect scaphoid
Click the "classidy and detect" button to load model and detect. Results 
are save to dicks and read againt to be cropped by Qt. It could be 
better to directly use results.imgs but I don't know how to use that to 
construct QPixmap yet....

The cropped region seems wrong. Need fix.